src/requirements.txt
dj-database-url==0.5.0
Django==2.0.4
django-crispy-forms==1.7.2
djangorestframework==3.8.1
gunicorn==19.7.1
olefile==0.44
Pillow==5.1.0
psycopg2==2.7.3
pytz==2017.2
